fix: isolate integration tests and harden EME interception
- Move device-dependent tests into the integration suite
- Rename the EME session initialization-data field
